+++ apm-agent-plugins/apm-redis-plugin/src/main/java/co/elastic/apm/agent/redis/jedis/package-info.java
@@ -1,0 +1,4 @@
+@NonnullApi
+package co.elastic.apm.agent.redis.jedis;
+
+import co.elastic.apm.agent.annotation.NonnullApi;